Split slow CLI tests for Vitest workspace
